% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expit}
\alias{expit}
\title{inverse logit transformation}
\usage{
expit(x)
}
\arguments{
\item{x}{quantity to be transformed}
}
\value{
real valued quantities
}
\description{
inverse logit transformation
}
\author{
Paul Conn \email{paul.conn@noaa.gov}
}
\keyword{expit}
\keyword{logit,}

