% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gIVH}
\alias{gIVH}
\title{a generic function to calculate the generalized independent variable hull for spatial models when
model is of form Data ~ f(g^{-1}(mu)), where f is a probability density/mass function, g is a link function, and
mu = X*Beta + K*Theta + Epsilon.  Here, X is a design matrix and Beta give fixed effects,
K is an expansion matrix for spatial (or GAM) random effects, Theta are random effects (if modeled), and Epsilon is normally distributed error}
\usage{
gIVH(X, Beta, Sampled, K = NULL, Theta = NULL, link = "log")
}
\arguments{
\item{X}{A design matrix for fixed effects (including sampled and unsampled locations)}

\item{Beta}{A matrix holding a sample from the joint (prior or posterior) distribution of regression parameters. The number of rows = the number of samples, while the number of columns gives the number of regression coefficients.}

\item{Sampled}{An integer-valued vector indicating which cells were actually sampled (these should correspond to rows of X)}

\item{K}{An expansion matrix for random effects (if not provided, a diagonal matrix is used)}

\item{Theta}{If random effects, a matrix holding a sample from the joint distribution of spatial random random effects (# rows = sample size, # columns = # of random effects).  Default is NULL.}

\item{link}{A character string providing the link function.  Options are "log", "logit," or "probit" (default = 'log').}
}
\value{
A list containing two objects: gIVH- a binary vector indicating which cells are in (=1.0) or are outside (=0.0) the gIVH, and Pred.var- a vector housing expected prediction variance.  Note that Pred.var is on the real scale, but does not incorporate stochasticity from the distribution f().
}
\description{
a generic function to calculate the generalized independent variable hull for spatial models when
model is of form Data ~ f(g^{-1}(mu)), where f is a probability density/mass function, g is a link function, and
mu = X*Beta + K*Theta + Epsilon.  Here, X is a design matrix and Beta give fixed effects,
K is an expansion matrix for spatial (or GAM) random effects, Theta are random effects (if modeled), and Epsilon is normally distributed error
}
\author{
Paul B. Conn
}
\keyword{simulation}
\keyword{spatio-temporal,}

