% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logit}
\alias{logit}
\title{logit transformation}
\usage{
logit(x)
}
\arguments{
\item{x}{quantity to be transformed}
}
\value{
logit value(s)
}
\description{
logit transformation
}
\author{
Paul Conn \email{paul.conn@noaa.gov}
}
\keyword{logit}

