% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_N_map}
\alias{plot_N_map}
\title{Function to plot abundance map for Bering Sea survey grid}
\usage{
plot_N_map(cur.t, N, Grid, highlight = NULL, cell.width = 1,
  leg.title = "Abundance", hcolor = "yellow", cur.max = NULL,
  rainbow = F)
}
\arguments{
\item{cur.t}{Time step to plot map for}

\item{N}{A matrix holding abundance point estimates; different rows correspond to different sampling units, columns correspond to time step}

\item{Grid}{A list object, each element of which holds a spatial polygons data frame for each time step}

\item{highlight}{If provided, this vector specifies which cells to separately highlight during plotting}

\item{cell.width}{if highlight is provided, cell.width must provide the width of a composite grid cell}

\item{leg.title}{Title for legend of plot (if different from the default "Abundance")}

\item{hcolor}{If highight provided, gives color for highlighting (default = "yellow")}

\item{cur.max}{If provided, sets maximum value for plotted abundance}

\item{rainbow}{If TRUE, use rainbow plot; otherwise use color blind safe palette}
}
\value{
An abundance map
}
\description{
Function to plot abundance map for Bering Sea survey grid
}
\author{
Paul Conn \email{paul.conn@noaa.gov}
}
\keyword{abundance}
\keyword{map,}
\keyword{plot}

