% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rrw}
\alias{rrw}
\title{SIMULATE AN ICAR PROCESS}
\usage{
rrw(Q)
}
\arguments{
\item{Q}{Precision matrix for the ICAR process}
}
\value{
Spatial random effects
}
\description{
SIMULATE AN ICAR PROCESS
}
\author{
Devin Johnson
}
\keyword{ICAR,}
\keyword{simulation}

