% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_data_generic}
\alias{sim_data_generic}
\title{function to simulate abundance over a simulated landscape}
\usage{
sim_data_generic(S, n.covs = 4, tau.epsilon = 100)
}
\arguments{
\item{S}{Number of cells (must be a square number)}

\item{n.covs}{Number of habitat covariates (default 4)}

\item{tau.epsilon}{Precision of exchangeable errors in log space}
}
\value{
A SpatialPolygonsDataFrame holding covariates, abundance, and lambda (expected abundance)
}
\description{
function to simulate abundance over a simulated landscape
}
\author{
Paul B. Conn
}
\keyword{simulation}
\keyword{spatio-temporal,}

