% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_effort}
\alias{sim_effort}
\title{function to simulate spatial count data over a simulated grid, conditional on expected abundance}
\usage{
sim_effort(Data, S, type = "random", prop.sampled = 0.1,
  n.points = round(0.1 * S))
}
\arguments{
\item{Data}{A SpatialPolygonsDataFrame for a square survey grid including a data column "N" that gives total abundance in each survey unit}

\item{S}{Number of cells (must be a square number)}

\item{type}{Type of spatial survey.  Choices are "random" (default), "clustered," or "balanced"}

\item{prop.sampled}{A numeric value giving the proportion of each survey unit that is actually sampled (default 0.1)}

\item{n.points}{Number of survey units that are sampled.  The default is to round (0.1*S)}
}
\value{
A data.frame with two columns, "Mapping" - indicating which survey units are sampled, and "Counts" - giving no. of animals counted
}
\description{
function to simulate spatial count data over a simulated grid, conditional on expected abundance
}
\author{
Paul B. Conn
}
\keyword{simulation}
\keyword{spatio-temporal,}

