% Generated by roxygen2 (4.0.2): do not edit by hand
\name{spat_pred}
\alias{spat_pred}
\title{function to perform Bayesian analysis of log Gaussian Cox processes and to make spatial predictions}
\usage{
spat_pred(formula, Data, Effort, spat.mod = 0, Offset, Area.adjust, Control,
  Assoc = NULL, K = NULL, Names.gam = NULL, Knots.gam = NULL,
  Prior.pars = NULL, Precision.pars = NULL)
}
\arguments{
\item{formula}{A formula object specifying the model for the fixed effects}

\item{Data}{A SpatialPolygonsDataFrame holding covariate values for all cells one wants to make predictions on (including sampled cells)}

\item{Effort}{A list including "Counts" which gives the response variable for surveyed cells, and "Mapping" which is a vector indicating which sample unit is associated with which count}

\item{spat.mod}{Controls whether (and what type) of spatial autocorrelation is implemented (spatmod = 0: no spatial autocorrelation; spatmod=1: process convolution; spatmod=2: srr ICAR)
Note that if spatmod=1, the argument Knots needs to be specified; if spatmod=2, Adj must be specified}

\item{Offset}{A vector giving the fraction of each grid cell in Mapping that was surveyed}

\item{Area.adjust}{A vector allowing for differences in suitable habitat for each cell.  Can be used for different grid cell sizes or different
proportions of suitable habitat (e.g., 1.0 = 100 percent of habitat is suitable for the focal species)}

\item{Control}{A list giving MCMC controls and additional model options
"iter": number of MCMC iterations;
"burnin": number of MCMC burnin iterations;
   "thin": if specified, how many iterations to skip between recorded posterior samples;
"srr.tol": Threshold eigenvalue level for SRR; only eigenvectors with higher eigenvalues than srr.tol are included in SRR formulation (default is 0.5)
"predict" If TRUE (default), calculate posterior predictions across the entire grid
"MH.nu" A vector providing continuous Uniform half-range values for Metropolis-Hastings proposals
"adapt" If true, adapts tuning parameters for nu updates; e.g., one could run a chain with adapt=TRUE, and use adapted MH tuning parameters in a longer analysis (default FALSE)
"fix.tau.epsilon" If TRUE, fixes tau.epsilon to 100
"Kern.gam.se" A vector setting gam se for kernels (length is number of smooth params)}

\item{Assoc}{Association matrix for all grid cells in Data (needed for ICAR models only)}

\item{K}{A (# cells)x(# knots) matrix - the 'K' matrix for process convolution modeling.}

\item{Names.gam}{A character vector giving the column names of Data that will be modeled as smooth effects}

\item{Knots.gam}{A list vector (length is # of smooth params), where each list element is a vector providing locations of knots (one location for each knot)}

\item{Prior.pars}{An optional list giving prior parameters; these include
"beta.tau" precision for Gaussian prior on regression parameters (default 0.1)
"a.eps" alpha parameter for tau_epsilon~Gamma(alpha,beta) (default = 1.0)
"b.eps" beta parameter for tau_epsilon~Gamma(alpha,beta) (default = 0.01)
 If provided, all values for Prior.pars need to be specified}

\item{Precision.pars}{If provided, this list gives tau.epsilon and tau.eta, and these parameters are both fixed during estimation}
}
\value{
returns a list with the following objecs:
	MCMC: A list object containing posterior samples;
 Accept: A list object indicating the number of proposals that were accepted for parameters updated via Metropolis-Hastings;
 Control: A list object giving MCMC tuning parameters (which are updated if the 'adapt' alorithm is used)
 gIVH: A binary vector, with length equal to the number of survey units, describing whether predictions for each survey unit are (=1) or are not (=0) within the gIVH
 K.gam: A ``design matrix" associated with smooth effects, should the GAM procedure be employed
}
\description{
function to perform Bayesian analysis of log Gaussian Cox processes and to make spatial predictions
}
\examples{
print("Later!")
}
\author{
Paul B. Conn \email{paul.conn@noaa.gov}
}
\keyword{areal}
\keyword{augmentation,}
\keyword{data}
\keyword{mcmc,}
\keyword{model,}
\keyword{prediction}
\keyword{spatial}

